pyproject.toml
[project]
name = "qa_service"
version = "0.1.0"
description = ""
authors = [
    {name = "KostylevEugene",email = "kostylev_evgenij@list.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (==2.0.40)",
    "alembic (==1.15.2)",
    "psycopg2-binary (==2.9.10)",
    "asyncpg (==0.30.0)",
    "ruff (>=0.12.4,<0.13.0)",
    "mypy (>=1.17.0,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "loguru (==0.7.2)",
    "httpx (>=0.28.1,<0.29.0)"
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
